Moved relevant files under TeachersPage/Chatbox/
