+++ src/components/pages/TeachersPage/DisplayOfChats.tsx
@@ -2,7 +2,7 @@ import { Grid } from '@mui/material';
 import { Dispatch, SetStateAction } from 'react';
 
 import { Student, StudentChat, SoloChat } from './types';
-import ReadOnlyChatbox from './ReadOnlyChatbox';
+import ReadOnlyChatbox from './Chatbox';
 
 interface DisplayOfChatsProps {
   studentChats: (StudentChat | SoloChat)[];
